5일간 날씨 UI 구현
- 5일치 날씨 데이터 표시
- mapping 단계에서 날씨 반올림

+++ domain/src/main/java/com/devhoseong/domain/model/Weather.kt
@@ -25,8 +25,8 @@ data class HourlyForecast(
 
 data class DailyForecast(
     val timestamp: Long,
-    val minTemp: Double,
-    val maxTemp: Double,
+    val minTemp: Int,
+    val maxTemp: Int,
     val weatherIcon: String
 )
